+++ src/Symfony/Component/Security/Http/Logout/SessionLogoutHandler.php
@@ -14,11 +14,16 @@ namespace Symfony\Component\Security\Http\Logout;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
+use Symfony\Component\Security\Http\EventListener\SessionLogoutListener;
+
+trigger_deprecation('symfony/security-http', '5.4', 'The "%s" class is deprecated, use "%s" instead.', SessionLogoutHandler::class, SessionLogoutListener::class);
 
 /**
  * Handler for clearing invalidating the current session.
  *
  * @author Johannes M. Schmitt <schmittjoh@gmail.com>
+ *
+ * @deprecated since Symfony 5.4, use {@link SessionLogoutListener} instead
  */
 class SessionLogoutHandler implements LogoutHandlerInterface
 {